+++ build_utils/create_and_pack_executable.py
@@ -1,9 +1,8 @@
-import codecs
+import argparse
+import logging
 import os
 import platform
-import re
 import shutil
-import sys
 import tarfile
 import zipfile
 
@@ -11,56 +10,73 @@ from PyInstaller.__main__ import run as pyinstaller_run
 
 import PartSeg
 
-if len(sys.argv) == 2:
-    base_path = os.path.abspath(sys.argv[1])
-else:
-    base_path = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+logger = logging.getLogger("bundle build")
+logger.setLevel(logging.INFO)
 
-os.chdir(base_path)
-
-pyinstaller_run(["-y", "--debug=all", "launcher.spec"])
+SYSTEM_NAME_DICT = {"Linux": "linux", "Windows": "windows", "Darwin": "macos"}
 
 
-def read(*parts):
-    with codecs.open(os.path.join(base_path, *parts), "r") as fp:
-        return fp.read()
+def create_archive(working_dir):
+    os.makedirs(os.path.join(working_dir, "dist2"), exist_ok=True)
+    file_name = f"PartSeg-{PartSeg.__version__}-{SYSTEM_NAME_DICT[platform.system()]}"
+    if platform.system() == "Darwin":
+        arch_file = tarfile.open(os.path.join(working_dir, "dist2", f"{file_name}.tgz"), "w:gz")
+        arch_file.write = arch_file.add
+    else:
+        arch_file = zipfile.ZipFile(os.path.join(working_dir, "dist2", f"{file_name}.zip"), "w", zipfile.ZIP_DEFLATED)
+    return arch_file
 
 
-def find_version(*file_paths):
-    version_file = read(*file_paths)
-    version_match = re.search(r"^version = ['\"]([^'\"]*)['\"]", version_file, re.M)
-    if version_match:
-        return version_match.group(1)
-    raise RuntimeError("Unable to find version string.")
+def fix_qt_location(working_dir, dir_name):
+    if platform.system() == "Darwin" and os.path.exists(os.path.join(working_dir, "dist", dir_name, "PyQt5", "Qt")):
+        shutil.move(
+            os.path.join(working_dir, "dist", dir_name, "PyQt5", "Qt"),
+            os.path.join(working_dir, "dist", dir_name, "PyQt5", "Qt5"),
+        )
 
 
-version = find_version(os.path.join(os.path.dirname(PartSeg.__file__), "version.py"))
+def create_bundle(spec_path, working_dir):
+    pyinstaller_args = [
+        "-y",
+        "--debug=all",
+        spec_path,
+        "--distpath",
+        os.path.join(working_dir, "dist"),
+        "--workpath",
+        os.path.join(working_dir, "build"),
+    ]
+    logger.info("run PyInstaller" + " ".join(pyinstaller_args))
 
-name_dict = {"Linux": "linux", "Windows": "windows", "Darwin": "macos"}
+    pyinstaller_run(pyinstaller_args)
 
-system_name = name_dict[platform.system()]
 
-os.makedirs(os.path.join(base_path, "dist2"), exist_ok=True)
+def archive_build(working_dir, dir_name):
+    base_zip_path = os.path.join(working_dir, "dist")
 
-if platform.system() == "Darwin":
-    arch_file = tarfile.open(os.path.join(base_path, "dist2", f"PartSeg-{version}-{system_name}.tgz"), "w:gz")
-    arch_file.write = arch_file.add
-else:
-    arch_file = zipfile.ZipFile(
-        os.path.join(base_path, "dist2", f"PartSeg-{version}-{system_name}.zip"), "w", zipfile.ZIP_DEFLATED
+    with create_archive(working_dir) as arch_file:
+        for root, _dirs, files in os.walk(os.path.join(working_dir, "dist", dir_name), topdown=False, followlinks=True):
+            for file_name in files:
+                arch_file.write(
+                    os.path.join(root, file_name), os.path.relpath(os.path.join(root, file_name), base_zip_path)
+                )
+
+
+def main():
+    parser = argparse.ArgumentParser("PartSeg build")
+    parser.add_argument(
+        "--spec", default=os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))), "launcher.spec")
     )
-base_zip_path = os.path.join(base_path, "dist")
+    parser.add_argument("--working-dir", default=os.path.abspath(os.curdir))
 
-dir_name = "PartSeg.app" if platform.system() == "Darwin2" else "PartSeg"
+    args = parser.parse_args()
 
-if platform.system() == "Darwin" and os.path.exists(os.path.join(base_path, "dist", dir_name, "PyQt5", "Qt")):
-    shutil.move(
-        os.path.join(base_path, "dist", dir_name, "PyQt5", "Qt"),
-        os.path.join(base_path, "dist", dir_name, "PyQt5", "Qt5"),
-    )
+    create_bundle(args.spec, args.working_dir)
 
-for root, dirs, files in os.walk(os.path.join(base_path, "dist", dir_name), topdown=False, followlinks=True):
-    for file_name in files:
-        arch_file.write(os.path.join(root, file_name), os.path.relpath(os.path.join(root, file_name), base_zip_path))
+    dir_name = "PartSeg.app" if platform.system() == "Darwin2" else "PartSeg"
 
-arch_file.close()
+    fix_qt_location(args.working_dir, dir_name)
+    archive_build(args.working_dir, dir_name)
+
+
+if __name__ == "__main__":
+    main()